Separate cli and core
### Added
- Ability to import package
- Type definitions for package

### Changed
- Error messages more usefull
