+++ src/cli/models/index.ts
@@ -1,14 +1,1 @@
-/**
- * @license
- * Copyright Andrey Chalkin <L2jLiga>. All Rights Reserved.
- *
- * Use of this source code is governed by an MIT-style license that can be
- * found in the LICENSE file at https://github.com/L2jLiga/xd2svg/LICENSE
- */
-
-export * from './artboard';
-export * from './artboard-info';
 export * from './cli-options';
-export * from './resource';
-export * from './shape';
-export * from './text';
