+++ src/core/lib/manifest-parser.ts
