+++ src/core/lib/resources-parser.ts
