+++ src/core/lib/styles/clip-path.ts
