+++ src/core/lib/styles/fill.ts
