+++ src/core/lib/styles/models/clip-path.ts
