+++ src/core/lib/styles/models/fill.ts
