+++ src/core/lib/styles/models/pattern.ts
