+++ src/core/lib/styles/models/stroke.ts
