+++ src/core/lib/styles/stroke.ts
