+++ src/core/lib/styles/text-attributes.ts
