+++ src/core/lib/svgo.ts
