+++ src/core/lib/utils/global-namespace.ts
