+++ src/core/models/artboard-info.ts
