+++ src/core/models/index.ts
@@ -1,4 +1,3 @@
-#!/usr/bin/env node
 /**
  * @license
  * Copyright Andrey Chalkin <L2jLiga>. All Rights Reserved.
@@ -6,8 +5,9 @@
  * Use of this source code is governed by an MIT-style license that can be
  * found in the LICENSE file at https://github.com/L2jLiga/xd2svg/LICENSE
  */
-'use strict';
 
-process.title = 'xd2svg';
-
-require('../dist/cli.js');
+export * from './artboard';
+export * from './artboard-info';
+export * from './resource';
+export * from './shape';
+export * from './text';
