+++ src/core/models/resource.ts
