+++ src/core/models/shape.ts
