+++ src/core/models/text.ts
