+++ src/xd2svg.ts
@@ -1,0 +1,1 @@
+export * from './core/xd2svg';
