+++ tsconfig.core.json
@@ -1,0 +1,10 @@
+{
+  "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "removeComments": false,
+    "declaration": true
+  },
+  "files": [
+    "src/cli.ts"
+  ]
+}
